changed README.md copy. moved files
