+++ scripts/countdown_erc721/README.md
@@ -5,11 +5,10 @@ This folder contains four scripts:
 - `1-deploy.ts`
 - `2-setOwner.ts`
 - `3-mintTo.ts`
-- `2-purchase.ts`
+- `4-purchase.ts`
 
 Below are the steps required to run the scripts successfully.
 
-</br>
 
 ## Prerequisites
 
@@ -18,8 +17,7 @@ Below are the steps required to run the scripts successfully.
 - Wallet private key
 - Network provider URL
 - Optional: Hardware wallet (e.g., Ledger)
-  
-</br>
+
 
 ## Environment Variables
 
@@ -28,9 +26,8 @@ Ensure the following environment variables are set:
 - `CUSTOM_ERC721_SALT`: Salt for deterministic address generation (must be a 32-character hexadecimal string)
 - `CUSTOM_ERC721_PROVIDER_URL`: Network provider URL
 - `HARDWARE_WALLET_ENABLED`: (Optional)  Flag indicating whether a hardware wallet is enabled
-- `HOLOGRAPH_ENVIRONMENT`: Holograph environment (localhost, develop, testnet, mainnet)
+- `HOLOGRAPH_ENVIRONMENT`: Holograph environment (develop, testnet, mainnet)
 
-</br>
 
 ## Script 1: Contract Deploy
 
@@ -47,6 +44,15 @@ This script utilizes hardcoded values, which need to be updated, to generate the
      - `contractName`: Name of the ERC721 contract
      - `contractSymbol`: Symbol of the ERC721 contract
      - `countdownERC721Initializer`: Initializer object containing various parameters for the ERC721 contract
+       - `description`: Description of the contract
+       - `imageURI`: Image URL for the NFT
+       - `externalLink`
+       - `encryptedMediaURI`
+       - `startDate`: Start date in seconds from epoc
+       - `initialMaxSupply`: Total Max Supply
+       - `mintIntervals`: How much to reduce counter by for a mint
+       - `contractURI`s
+         - `publicSalePrice`: Note that `10_000_000 = $10` USD
 
 3. **Preparing to Deploy Contract:**
    - Encode contract initialization parameters.
@@ -56,30 +62,27 @@ This script utilizes hardcoded values, which need to be updated, to generate the
    - Execute the deployment process.
    - Upon successful deployment, the contract address will be displayed.
 
-</br>
 
 To run the script, use the following command in your terminal:
 
 ```sh
-ts-node countdown_erc721/deploy.ts
+ts-node ./scripts/countdown_erc721/1-deploy.ts
 ```
 
-</br>
-
-## Script 2: Set Owner
+## Script 2: Set Owner [WIP]
 
 This is a straightforward script that invokes the contract to update the owner using hardcoded values.
 
 ### Steps
 
-2. **Set the static values:**
+1. **Set the static values:**
    - Configure the static values:
      - `contractAddress`: The CountdownERC721 address
      - `newOwner`: The wallet address of the new owner.
-  
-3. **Run the script update the owner:**
+2. **Run the script update the owner:**
+
 ```sh
-ts-node countdown_erc721/2-setOwner.ts 
+ts-node ./scripts/countdown_erc721/2-setOwner.ts 
 ```
 
 ## Script 3: Mint To
@@ -88,15 +91,16 @@ This is a straightforward script that invokes the contract `mintTo` function usi
 
 ### Steps
 
-2. **Set the static values:**
+1. **Set the static values:**
    - Configure the static values:
      - `contractAddress`: The CountdownERC721 address
      - `recipient`: The wallet address of recipient.
      - `quantity`: The number of tokens to be minted.
   
-3. **Run the script update the owner:**
+2. **Run the script update the owner:**
+
 ```sh
-ts-node countdown_erc721/3-mintTo.ts 
+ts-node ./scripts/countdown_erc721/3-mintTo.ts 
 ```
 
 ## Script 4: Purchase
@@ -105,13 +109,14 @@ This is a straightforward script that invokes the contract `purchase` function u
 
 ### Steps
 
-2. **Set the static values:**
+1. **Set the static values:**
    - Configure the static values:
      - `contractAddress`: The CountdownERC721 address
      - `quantity`: The number of tokens to be minted.
      - `price`: The public sale price in wei.
   
-3. **Run the script update the owner:**
+2. **Run the script update the owner:**
+
 ```sh
-ts-node countdown_erc721/4-purchase.ts 
+ts-node ./scripts/countdown_erc721/4-purchase.ts 
 ```
